(PC-15031) refactor(typography): move getTextAttrs and getHeadingAttrs into their own files
