+++ src/features/auth/signup/underageSignup/notificationPagesStyles.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components/native'
 
 import { getSpacing, Typo } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 export const StyledSubtitle = styled(Typo.Title4).attrs(getHeadingAttrs(undefined))({
   textAlign: 'center',
