+++ src/features/bookings/components/BookingDetailsTicketContent.tsx
@@ -13,7 +13,7 @@ import { ButtonWithLinearGradient } from 'ui/components/buttons/ButtonWithLinear
 import { TouchableLink } from 'ui/components/touchableLink/TouchableLink'
 import { ExternalSite } from 'ui/svg/icons/ExternalSite'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 export type BookingDetailsTicketContentProps = {
   booking: BookingReponse
