+++ src/features/offer/pages/OfferBody.tsx
@@ -25,7 +25,7 @@ import { Hero } from 'ui/components/hero/Hero'
 import { SectionWithDivider } from 'ui/components/SectionWithDivider'
 import { Flag as DefaultFlag } from 'ui/svg/icons/Flag'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 import { useOffer } from '../api/useOffer'
 import { OfferIconCaptions, OfferPartialDescription } from '../components'
