+++ src/features/profile/components/CreditCeiling.tsx
@@ -12,7 +12,7 @@ import { Offers } from 'ui/svg/icons/Offers'
 import { getSpacing, Typo } from 'ui/theme'
 // eslint-disable-next-line no-restricted-imports
 import { ColorsEnum } from 'ui/theme/colors'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 import { AnimatedProgressBar } from '../../../ui/components/bars/AnimatedProgressBar'
