+++ src/features/profile/pages/DeleteProfile/ConfirmDeleteProfile.tsx
@@ -17,7 +17,7 @@ import { BicolorError } from 'ui/svg/icons/BicolorError'
 import { Dot } from 'ui/svg/icons/Dot'
 import { ExternalSiteFilled } from 'ui/svg/icons/ExternalSiteFilled'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 import { Li } from 'ui/web/list/Li'
 import { VerticalUl } from 'ui/web/list/Ul'
