+++ src/features/venue/pages/VenueBody.tsx
@@ -18,7 +18,7 @@ import { Hero } from 'ui/components/hero/Hero'
 import { SectionWithDivider } from 'ui/components/SectionWithDivider'
 import { LocationPointer as DefaultLocationPointer } from 'ui/svg/icons/LocationPointer'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 import { useVenue } from '../api/useVenue'
 import { VenuePartialAccordionDescription } from '../components/VenuePartialAccordionDescription'
