+++ src/libs/geolocation/components/WhereSection.tsx
@@ -15,7 +15,7 @@ import { TouchableLink } from 'ui/components/touchableLink/TouchableLink'
 import { ArrowNext as DefaultArrowNext } from 'ui/svg/icons/ArrowNext'
 import { BicolorLocationBuilding as LocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding'
 import { Typo, getSpacing } from 'ui/theme'
-import { getHeadingAttrs } from 'ui/theme/typography'
+import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
 type Props = {
   beforeNavigateToItinerary?: () => Promise<void> | void
