+++ src/ui/components/InputLabel/InputLabel.web.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { getTextAttrs } from 'ui/theme/typography'
+import { getTextAttrs } from 'ui/theme/typographyAttrs/getTextAttrs'
 
 export const InputLabel = styled.label.attrs(getTextAttrs())(({ theme }) => ({
   ...theme.typography?.body,
