+++ src/ui/theme/typographyAttrs/getTextAttrs.ts
@@ -1,0 +1,3 @@
+import { Platform } from 'react-native'
+
+export const getTextAttrs = (dir?: string) => (Platform.OS === 'web' ? { dir: dir ?? 'ltr' } : {})
